rename remark plugins
