--- packages/remark-embed-plus/src/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./lib/remark-embed-plus";
