+++ packages/remark-embed/README.md
@@ -1,3 +1,3 @@
-# remark-embed-plus
+# @flowershow/remark-embed
 
 Converts Youtube link surrounded by newlines in markdown to embedded iframe
