+++ packages/remark-embed/package.json
@@ -1,11 +1,11 @@
 {
-  "name": "remark-embed-plus",
+  "name": "@flowershow/remark-embed",
   "version": "1.0.2",
   "description": "Converts youtube link in mdx to an iframe embed",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/flowershow/flowershow.git",
-    "directory": "packages/remark-embed-plus"
+    "directory": "packages/remark-embed"
   },
   "keywords": [
     "remark",
