+++ packages/remark-embed/project.json
@@ -1,21 +1,21 @@
 {
-  "name": "remark-embed-plus",
+  "name": "remark-embed",
   "$schema": "../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "packages/remark-embed-plus/src",
+  "sourceRoot": "packages/remark-embed/src",
   "projectType": "library",
   "targets": {
     "lint": {
       "executor": "@nrwl/linter:eslint",
       "outputs": ["{options.outputFile}"],
       "options": {
-        "lintFilePatterns": ["packages/remark-embed-plus/**/*.ts"]
+        "lintFilePatterns": ["packages/remark-embed/**/*.ts"]
       }
     },
     "test": {
       "executor": "@nrwl/jest:jest",
       "outputs": ["{workspaceRoot}/coverage/{projectRoot}"],
       "options": {
-        "jestConfig": "packages/remark-embed-plus/jest.config.ts",
+        "jestConfig": "packages/remark-embed/jest.config.ts",
         "passWithNoTests": true
       }
     },
@@ -23,17 +23,17 @@
       "executor": "@nrwl/web:rollup",
       "outputs": ["{options.outputPath}"],
       "options": {
-        "entryFile": "packages/remark-embed-plus/src/index.ts",
-        "outputPath": "packages/remark-embed-plus/dist",
+        "entryFile": "packages/remark-embed/src/index.ts",
+        "outputPath": "packages/remark-embed/dist",
         "compiler": "babel",
-        "tsConfig": "packages/remark-embed-plus/tsconfig.lib.json",
-        "project": "packages/remark-embed-plus/package.json",
+        "tsConfig": "packages/remark-embed/tsconfig.lib.json",
+        "project": "packages/remark-embed/package.json",
         "format": ["esm", "cjs"],
         "external": ["unist-util-visit"],
         "generateExportsField": true,
         "assets": [
           {
-            "glob": "packages/remark-embed-plus/README.md",
+            "glob": "packages/remark-embed/README.md",
             "input": ".",
             "output": "."
           }
