+++ packages/remark-embed/src/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./lib/remark-embed";
