--- packages/remark-wiki-link-plus/src/index.js
@@ -1,2 +1,0 @@
-export * from "./lib/remark-wiki-link-plus.js";
-export { default } from "./lib/remark-wiki-link-plus.js";
