+++ packages/remark-wiki-link/package.json
@@ -1,11 +1,11 @@
 {
-  "name": "remark-wiki-link-plus",
+  "name": "@flowershow/remark-wiki-link",
   "version": "1.1.0",
   "description": "Parse and render wiki-style links in markdown especially Obsidian style links.",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/flowershow/flowershow.git",
-    "directory": "packages/remark-wiki-link-plus"
+    "directory": "packages/remark-wiki-link"
   },
   "keywords": [
     "remark",
