+++ packages/remark-wiki-link/src/lib/getFiles.js
