+++ packages/template/styles/global.css
@@ -1,4 +1,4 @@
-@import "remark-callouts/styles.css";
+@import "@flowershow/remark-callouts/styles.css";
 
 /* mathjax */
 .math-inline {
